Moving to exp tree
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@9288 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ exp/include/Galois/Runtime/WorkListAlt.h
@@ -139,7 +139,7 @@ public:
     
     //steal
     int id = local.myEffectiveID();
-    for (int i = 0; i < local.size(); ++i) {
+    for (int i = 0; i < (int) local.size(); ++i) {
       ++id;
       id %= local.size();
       ret = me.steal(local.get(id));
